+++ app/client/cypress/e2e/Regression/ClientSide/Widgets/TableV2/TableV2_PropertyPane_2_spec.js
@@ -1,11 +1,10 @@
 import {
-  entityExplorer,
-  table,
-  propPane,
   agHelper,
   deployMode,
   draggableWidgets,
   locators,
+  propPane,
+  table,
 } from "../../../../../support/Objects/ObjectsCore";
 const widgetsPage = require("../../../../../locators/Widgets.json");
 
@@ -114,6 +113,11 @@ describe(
     it("4. It provides currentRow and currentIndex properties in min validation field", function () {
       agHelper.AddDsl("tableV2NewDslWithPagination");
       cy.openPropertyPane("tablewidgetv2");
+      // To make sorting more predictable making `id` column type Number.
+      table.toggleColumnEditableViaColSettingsPane("id", "v2", true, false);
+      cy.changeColumnType("Number");
+
+      propPane.NavigateBackToPropertyPane();
       table.toggleColumnEditableViaColSettingsPane(
         "orderAmount",
         "v2",
@@ -124,17 +128,18 @@ describe(
       propPane.UpdatePropertyFieldValue("Computed value", "{{currentIndex}}");
       cy.changeColumnType("Number");
 
+      // #region Min
       propPane.UpdatePropertyFieldValue("Min", "{{currentIndex}}");
       cy.get(".t--evaluatedPopup-error").should("not.exist");
 
       // Update cell with row : 1, column : orderAmount
 
       table.EditTableCell(1, 4, -1, false);
-      cy.get(".bp3-popover-content").contains("Invalid input");
+      cy.get(locators._popoverToolTip).contains("Invalid input");
       table.UpdateTableCell(1, 4, 0);
-      cy.get(".bp3-popover-content").contains("Invalid input");
+      cy.get(locators._popoverToolTip).contains("Invalid input");
       table.UpdateTableCell(1, 4, 3);
-      cy.get(".bp3-popover-content").should("not.exist");
+      cy.get(locators._popoverToolTip).should("not.exist");
 
       // Check if currentRow works
       propPane.NavigateBackToPropertyPane();
@@ -148,25 +153,39 @@ describe(
 
       // Update cell with row : 0, column : orderAmount. The min is set to 7 (i.e value of cell in id column)
       table.EditTableCell(1, 4, 8, false);
-      cy.get(".bp3-popover-content").should("not.exist");
+      cy.get(locators._popoverToolTip).should("not.exist");
 
-      table.UpdateTableCell(1, 4, 6);
-      cy.get(".bp3-popover-content").contains("Row at index 1 is not valid");
+      table.UpdateTableCell(1, 4, -1);
+      cy.get(locators._popoverToolTip).contains("Row at index 1 is not valid");
 
       table.UpdateTableCell(1, 4, 8);
-      cy.get(".bp3-popover-content").should("not.exist");
+      cy.get(locators._popoverToolTip).should("not.exist");
 
       propPane.UpdatePropertyFieldValue(
         "Error message",
         "Row with id {{currentRow.id}} is not valid",
       );
 
-      table.EditTableCell(1, 4, 5, false);
-      cy.get(".bp3-popover-content").contains("Row with id 7 is not valid");
+      table.EditTableCell(1, 4, 1, false);
+      cy.get(locators._popoverToolTip).contains("Row with id 2 is not valid");
+
+      cy.discardTableCellValue(4, 1);
+      table.DiscardEditRow(1, 5);
+
+      // Sort the id column and ensure that `currentRow` is correctly evaluated
+      table.SortColumn("id", "descending");
+      table.EditTableCell(1, 4, 1, false);
+      cy.get(locators._popoverToolTip).contains("Row with id 24 is not valid");
+
+      cy.discardTableCellValue(4, 1);
+      table.SortColumn("id", "ascending");
 
       propPane.UpdatePropertyFieldValue("Min", "");
       propPane.UpdatePropertyFieldValue("Error message", "");
 
+      // #endregion Min
+
+      // #region Regex
       // Check for currentIndex property on Regex field
       cy.changeColumnType("Plain text");
       propPane.UpdatePropertyFieldValue("Regex", "{{currentIndex}}2");
@@ -174,68 +193,101 @@ describe(
       cy.get(".t--evaluatedPopup-error").should("not.exist");
       table.EditTableCell(1, 4, 3, false);
 
-      cy.get(".bp3-popover-content").contains("Invalid input");
+      cy.get(locators._popoverToolTip).contains("Invalid input");
       table.UpdateTableCell(1, 4, 12);
-      cy.get(".bp3-popover-content").should("not.exist");
+      cy.get(locators._popoverToolTip).should("not.exist");
 
       // Check for currentRow property on Regex field
       propPane.UpdatePropertyFieldValue("Regex", "{{currentRow.id}}");
-      table.EditTableCell(1, 4, 7, false);
-      cy.get(".bp3-popover-content").should("not.exist");
+      table.EditTableCell(1, 4, 20, false);
+      cy.get(locators._popoverToolTip).should("not.exist");
       table.UpdateTableCell(1, 4, 8);
-      cy.get(".bp3-popover-content").contains("Invalid input");
-      table.UpdateTableCell(1, 4, 7);
-      cy.get(".bp3-popover-content").should("not.exist");
+      cy.get(locators._popoverToolTip).contains("Invalid input");
+      table.UpdateTableCell(1, 4, 20);
+      cy.get(locators._popoverToolTip).should("not.exist");
+
+      cy.discardTableCellValue(4, 1);
+      table.DiscardEditRow(1, 5);
+
+      // Sort the id column and ensure that `currentRow` is correctly evaluated
+      table.SortColumn("id", "descending");
+
+      table.EditTableCell(1, 4, 10, false);
+      cy.get(locators._popoverToolTip).contains("Invalid input");
+
+      table.UpdateTableCell(1, 4, 24);
+      cy.get(locators._popoverToolTip).should("not.exist");
+
       propPane.UpdatePropertyFieldValue("Regex", "");
 
+      table.DiscardEditRow(1, 5);
+      table.SortColumn("id", "ascending");
+
+      // #endregion Regex
+
+      // #region Required
       propPane.EnterJSContext("Required", "{{currentIndex == 1}}");
       table.EditTableCell(1, 4, "", false);
-      cy.get(".bp3-popover-content").contains("This field is required");
+      cy.get(locators._popoverToolTip).contains(
+        Cypress.env("MESSAGES").FIELD_REQUIRED_MESSAGE(),
+      );
       table.UpdateTableCell(1, 4, 1, true);
-      cy.get(".bp3-popover-content").should("not.exist");
-
-      cy.wait(1500);
-      cy.discardTableRow(5, 1);
-      cy.wait(1500);
+      cy.get(locators._popoverToolTip).should("not.exist");
 
       // Value isn't required in Row Index 2
       table.EditTableCell(2, 4, "", false);
-      cy.get(".bp3-popover-content").should("not.exist");
+      cy.get(locators._popoverToolTip).should("not.exist");
       table.UpdateTableCell(2, 4, "11");
-      cy.get(".bp3-popover-content").should("not.exist");
+      cy.get(locators._popoverToolTip).should("not.exist");
       table.UpdateTableCell(2, 4, "", true);
-      cy.get(".bp3-popover-content").should("not.exist");
+      cy.get(locators._popoverToolTip).should("not.exist");
 
-      cy.wait(1500);
-      cy.discardTableRow(5, 2);
-
-      // Check for Required property using currentRow, row with index 1 has id 7
-      propPane.UpdatePropertyFieldValue("Required", "{{currentRow.id == 7}}");
+      table.DiscardEditRow(2, 5);
+      // Check for Required property using currentRow, row with index 1 has id 2
+      propPane.UpdatePropertyFieldValue("Required", "{{currentRow.id == 2}}");
 
       table.EditTableCell(1, 4, "", false);
-      cy.get(".bp3-popover-content").contains("This field is required");
+      cy.get(locators._popoverToolTip).contains(
+        Cypress.env("MESSAGES").FIELD_REQUIRED_MESSAGE(),
+      );
       table.UpdateTableCell(1, 4, 1);
-      cy.get(".bp3-popover-content").should("not.exist");
+      cy.get(locators._popoverToolTip).should("not.exist");
       table.UpdateTableCell(1, 4, "");
-      cy.get(".bp3-popover-content").contains("This field is required");
+      cy.get(locators._popoverToolTip).contains(
+        Cypress.env("MESSAGES").FIELD_REQUIRED_MESSAGE(),
+      );
 
       table.UpdateTableCell(1, 4, "1", true);
-      cy.get(".bp3-popover-content").should("not.exist");
-
-      cy.wait(1500);
-      cy.discardTableRow(5, 1);
-      cy.wait(1500);
+      cy.get(locators._popoverToolTip).should("not.exist");
 
       // Value isn't required in Row Index 2
       table.EditTableCell(2, 4, "", false);
-      cy.get(".bp3-popover-content").should("not.exist");
+      cy.get(locators._popoverToolTip).should("not.exist");
       table.UpdateTableCell(2, 4, 10);
-      cy.get(".bp3-popover-content").should("not.exist");
+      cy.get(locators._popoverToolTip).should("not.exist");
       table.UpdateTableCell(2, 4, "", true);
-      cy.get(".bp3-popover-content").should("not.exist");
+      cy.get(locators._popoverToolTip).should("not.exist");
 
-      cy.wait(1500);
-      cy.discardTableRow(5, 2);
+      table.DiscardEditRow(2, 5);
+
+      // Sort the id column and ensure that `currentRow` is correctly evaluated
+      table.SortColumn("id", "descending");
+      table.EditTableCell(1, 4, "", false);
+      cy.get(locators._popoverToolTip).should("not.exist");
+      table.UpdateTableCell(1, 4, 10);
+      cy.get(locators._popoverToolTip).should("not.exist");
+
+      propPane.UpdatePropertyFieldValue("Required", "{{currentRow.id == 24}}");
+      table.EditTableCell(1, 4, "");
+      cy.get(locators._popoverToolTip).contains(
+        Cypress.env("MESSAGES").FIELD_REQUIRED_MESSAGE(),
+      );
+
+      cy.discardTableCellValue(4, 1);
+      table.DiscardEditRow(1, 5);
+
+      table.SortColumn("id", "ascending");
+      // #endregion Required
 
       // Cleanup
       propPane.UpdatePropertyFieldValue(
