+++ app/client/src/ce/constants/messages.ts
@@ -2500,3 +2500,5 @@ export const PAGE_ENTITY_NAME = "Page";
 
 export const EMPTY_DATASOURCE_TOOLTIP_SIDEBUTTON = () =>
   "Create a datasource to power your app with data.";
+
+export const FIELD_REQUIRED_MESSAGE = () => `This field is required`;
