+++ app/client/src/components/propertyControls/TableInlineEditValidPropertyControl.tsx
@@ -28,7 +28,7 @@ const getBindingSuffix = (tableId: string, columnName: string) => {
     ))
     (
       (${tableId}.isAddRowInProgress ? ${tableId}.newRow.${columnName} : ${tableId}.columnEditableCellValue.${columnName}) || "",
-      ${tableId}.isAddRowInProgress ? ${tableId}.newRow : (${tableId}.processedTableData[${tableId}.editableCell.index] ||
+      ${tableId}.isAddRowInProgress ? ${tableId}.newRow : (${tableId}.processedTableData[${tableId}.editableCell.${ORIGINAL_INDEX_KEY}] ||
         Object.keys(${tableId}.processedTableData[0])
           .filter(key => ["${ORIGINAL_INDEX_KEY}", "${PRIMARY_COLUMN_KEY_VALUE}"].indexOf(key) === -1)
           .reduce((prev, curr) => {
